Move CachePolicySuite tests back to SQLSuite.
Also, gave clearer names to partition-caching table properties.
